+++ jakarta-ee9-webapp-shopping-cart-filters-database/src/main/java/com/kevinpina/model/ItemCart.java
@@ -12,7 +12,7 @@ public class ItemCart {
 	private Integer quantity;
 	private Product product;
 
-	public float getAmount() {
+	public Double getAmount() {
 		return quantity * product.getPrice();
 	}
