+++ jakarta-ee9-webapp-shopping-cart-filters-database/src/main/java/com/kevinpina/model/Product.java
@@ -13,7 +13,7 @@ public class Product {
 
 	private Long id;
 	private String name;
-	private Float price;
+	private Double price;
 	private String categoryName;
 	
 	/**
